requirements.txt
Flask
numpy
tensorflow
pandas
Pillow